app/serializers.py
# backend/serializers.py
from rest_framework import serializers
from .models import Account, Room


class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = Account
        fields = "__all__"


class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = "__all__"
